My second commit
I have created a ballbounce class which creates an array of balls and bounces them around the screen. On new game, I am now using the line class which moves according to the location of the mouse. I have created a new driver class - this runs the game. Updates to CreateBricks class.